requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
urllib3==2.1.0
pandas==2.1.3
plotly==5.18.0
textstat==0.7.3
nltk==3.8.1
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==3.0.1

